+++ sdk/keyvault/azure_security_keyvault_keys/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Breaking Changes
 
+- Changed model fields defined as an `Option<HashMap<K, V>>` to just a `HashMap<K, V>`.
+- Changed model fields defined as an `Option<Vec<T>>` to just a `Vec<T>`.
+- Renamed `KeyClient::get_keys` to `list_keys`.
+
 ### Bugs Fixed
 
 ### Other Changes